Eh.
To really make this work like struct, and to better fit the model of
what tuples and lists ought to mean, unpack() should return a tuple. But
that slows it down, except here in the EVT_RESOLUTION_INFO section. I
could speed it up again by removing the list from unpack() and going
straight to a tuple, but that would be incompatible with older versions
of Python.